requirements.txt
folium==0.17.0
HTMLTestRunner-rv==1.1.2
matplotlib==3.9.0
numpy==2.0.0
pandas==2.2.2
seaborn==0.13.2
selenium==4.22.0
webdriver-manager==4.0.1
pillow~=10.3.0